refactor query forms
