+++ daiquiri_core/static/core/css/style.scss
@@ -136,3 +136,7 @@ div.control-label {
 .daiquiri-multi-checkbox > :last-child {
     margin-bottom: 0;
 }
+
+div.fieldset {
+    margin-bottom: 20px;
+}
