+++ daiquiri_query/static/query/js/forms/box.js
@@ -1,0 +1,17 @@
+app.factory('BoxFormService', ['QueryService', function(QueryService) {
+
+    var service = {
+        values: {},
+        errors: {}
+    };
+
+    service.activate = function() {
+        QueryService.activateForm('box');
+    };
+
+    service.submit = function() {
+
+    };
+
+    return service;
+}]);
