+++ daiquiri_query/static/query/js/forms/cone.js
@@ -1,0 +1,17 @@
+app.factory('ConeFormService', ['QueryService', function(QueryService) {
+
+    var service = {
+        values: {},
+        errors: {}
+    };
+
+    service.activate = function() {
+        QueryService.activateForm('cone');
+    };
+
+    service.submit = function() {
+
+    };
+
+    return service;
+}]);
