+++ daiquiri_query/templates/query/query_form_cone.html
@@ -1,5 +1,56 @@
 {% load i18n %}
 
+<p>
+    {% trans 'Please specify a coordinate in right ascension (RA) and declination (DEC) and a cone radius. The query will result in objects that have coordinates in the search cone.' %}
+</p>
+
+<div class="row fieldset">
+    <div class="col-md-4">
+        <formgroup
+            data-id="ra"
+            data-type="text"
+            data-label="{% trans 'RA (degrees)' %}"
+            data-model="service.values.cone.ra"
+            data-errors="service.errors.cone.ra">
+        </formgroup>
+    </div>
+    <div class="col-md-4">
+        <formgroup
+            data-id="de"
+            data-type="text"
+            data-label="{% trans 'DE (degrees)' %}"
+            data-model="service.values.cone.de"
+            data-errors="service.errors.cone.de">
+        </formgroup>
+    </div>
+    <div class="col-md-4">
+        <formgroup
+            data-id="radius"
+            data-type="text"
+            data-label="{% trans 'Radius (arcsec)' %}"
+            data-model="service.values.cone.radius"
+            data-errors="service.errors.cone.radius">
+        </formgroup>
+    </div>
+</div>
+
+<formgroup
+    data-id="table_name"
+    data-type="text"
+    data-label="{% trans 'Table name' %}"
+    data-model="service.values.sql.table_name"
+    data-errors="service.errors.sql.table_name">
+</formgroup>
+
+<button class="btn btn-primary" ng-click="service.submitJob('sql')">
+    {% trans 'Submit new SQL Query' %}
+</button>
+
+<button class="btn" ng-click="">
+    {% trans 'Clear input window' %}
+</button>
+{% load i18n %}
+
