+++ daiquiri_query/templates/query/query_form_sql.html
@@ -1,5 +1,9 @@
 {% load i18n %}
 
+<p>
+   {% trans 'Place your SQL statement directly in the text area below and submit your request using the button.' %}
+</p>
+
 <div class="btn-group daiquiri-query-form-bar" role="group">
     <button type="button" class="btn btn-default">Databases</button>
     <button type="button" class="btn btn-default">Functions</button>
@@ -10,27 +14,23 @@
     data-id="sql_query"
     data-type="codemirror"
     data-label="{% trans 'SQL query' %}"
-    data-model="service.values.sql.query"
-    data-errors="service.errors.sql.query.messages"
+    data-model="service.forms.sql.values.query"
+    data-errors="service.forms.sql.errors.query.messages"
     data-mode="text/x-mariadb">
 </formgroup>
 
 <formgroup
-    data-id="table_name"
+    data-id="sql_table_name"
     data-type="text"
     data-label="{% trans 'Table name' %}"
-    data-model="service.values.sql.table_name"
-    data-errors="service.errors.sql.table_name">
+    data-model="service.forms.sql.values.table_name"
+    data-errors="service.forms.sql.errors.table_name">
 </formgroup>
 
-<button class="btn btn-primary" ng-click="service.submitJob('sql')">
+<button class="btn btn-primary" ng-click="service.forms.sql.submit()">
     {% trans 'Submit new SQL Query' %}
 </button>
 
 <button class="btn" ng-click="">
     {% trans 'Clear input window' %}
 </button>
-
-<pre>
-select count(id from daiquiri_data_obs.stars
-</pre>
