+++ daiquiri_query/templates/query/query_sidebar.html
@@ -20,7 +20,7 @@ The guest user is using 28.1 MB of its quota of 100.0 MB.
 
         <li class="list-group-item list-group-link">
             <a href=""
-                ng-click="service.activateForm('{{ form.key }}')"
+                ng-click="service.forms['{{ form.key }}'].activate()"
                 ng-class="{active: service.form == '{{ form.key }}'}">
                 {{ form.label }}
             </a>
